backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
prometheus-client==0.19.0
